requirements.txt
numpy
opencv-python==4.8.1.78
scikit-learn==1.3.2
Flask==2.3.3
Pillow==10.0.0
joblib==1.3.2
Werkzeug==2.3.7